+++ fabric/src/main/java/wiki/minecraft/heywiki/fabric/TitleScreenInterface.java
@@ -1,0 +1,5 @@
+package wiki.minecraft.heywiki.fabric;
+
+public interface TitleScreenInterface {
+    void heywiki$resetInitialized();
+}