+++ CHANGELOG.md
@@ -1,5 +1,5 @@
-# v1.4.1+1.20.4
+# v1.4.2+1.20.4
 
 ## Fixes
 
-- Fixes a critical bug where the the mod attempts to open the wiki page every time something is being aimed at
+- Fix performance issues with TitleScreen (Fabric)